requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pytz==2023.3
pydantic==2.5.0
python-multipart==0.0.6